Remove component of point in normal direction of element plane
The spherical coordinate transformation of a point computes a cartesian
location that is outside the element plane. When computing barycentric
coordinates, this makes it possible for the sum of the barycentric
coordinates to exceed 1. To fix this, we remove the component of the
particle's cartesian coordinates that is normal to the element plane
when performing point in cell checks