Change UX->Ux interpolators prefix for unstructured grid interpolators
Note also that changes in test assertions were needed for unstructured
grid pset.execute tests that use spherical mesh types. This arises due
to the fact that we now alter the particle locations by removing the
component of the spherical coordinate transformation that is not in the
element plane; this changes the resulting barycentric coordinates and
thus interpolated values for particle advection.